Compatibility with new numpy

+++ neurolang/datalog/wrapped_collections.py
@@ -19,8 +19,8 @@ REBV = ReplaceExpressionsByValues(dict())
 
 class WrappedTypeMap:
     row_maps = {
-        np.integer: int,
-        np.float: float
+        np.int64: int,
+        np.float64: float
     }
 
     @lru_cache(maxsize=256)
